Copy FeatureLayerGeopackage samples as starting point for RasterLayerGeopackage
