+++ src/Android/Xamarin.Android/groups.json
@@ -247,8 +247,8 @@
               "Path": "Samples/Data/AddShapefile"
             },
             {
-              "SampleName": "FeatureLayerGeoPackage",
-              "Path": "Samples/Data/FeatureLayerGeoPackage"
+              "SampleName": "RasterLayerGeoPackage",
+              "Path": "Samples/Data/RasterLayerGeoPackage"
             }
           ]
         }
