+++ src/UWP/ArcGISRuntime.UWP.Samples/ArcGISRuntime.UWP.Samples.CSharp.csproj
@@ -85,7 +85,7 @@
   </PropertyGroup>
   <ItemGroup>
     <!-- A reference to the entire .Net Framework and Windows SDK are automatically included -->
-    <Content Include="Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.jpg">
+    <Content Include="Samples\Data\RasterLayerGeoPackage\RasterLayerGeoPackage.jpg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Samples\Data\AddShapefile\AddShapefile.jpg">
@@ -459,10 +459,10 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
-    <Content Include="Samples\Data\FeatureLayerGeoPackage\metadata.json">
+    <Content Include="Samples\Data\RasterLayerGeoPackage\metadata.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
-    <Page Include="Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.xaml">
+    <Page Include="Samples\Data\RasterLayerGeoPackage\RasterLayerGeoPackage.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
@@ -492,8 +492,8 @@
     <Compile Include="Samples\Data\GenerateGeodatabase\GenerateGeodatabase.xaml.cs">
       <DependentUpon>GenerateGeodatabase.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.xaml.cs">
-      <DependentUpon>FeatureLayerGeoPackage.xaml</DependentUpon>
+    <Compile Include="Samples\Data\RasterLayerGeoPackage\RasterLayerGeoPackage.xaml.cs">
+      <DependentUpon>RasterLayerGeoPackage.xaml</DependentUpon>
     </Compile>
     <Compile Include="Samples\Data\AddShapefile\AddShapefile.xaml.cs">
       <DependentUpon>AddShapefile.xaml</DependentUpon>
