+++ src/UWP/ArcGISRuntime.UWP.Samples/Samples/Data/RasterLayerGeoPackage/RasterLayerGeoPackage.jpg
