+++ src/UWP/ArcGISRuntime.UWP.Samples/Samples/Data/RasterLayerGeoPackage/RasterLayerGeoPackage.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl 
-    x:Class="ArcGISRuntime.UWP.Samples.FeatureLayerGeoPackage.FeatureLayerGeoPackage"
+    x:Class="ArcGISRuntime.UWP.Samples.RasterLayerGeoPackage.RasterLayerGeoPackage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:esriUI="using:Esri.ArcGISRuntime.UI.Controls" >
