+++ src/UWP/ArcGISRuntime.UWP.Samples/groups.json
@@ -243,8 +243,8 @@
               "Path": "Samples/Data/GenerateGeodatabase"
             },
             {
-              "SampleName": "FeatureLayerGeoPackage",
-              "Path": "Samples/Data/FeatureLayerGeoPackage"
+              "SampleName": "RasterLayerGeoPackage",
+              "Path": "Samples/Data/RasterLayerGeoPackage"
             },
             {
               "SampleName": "AddShapefile",
