+++ src/WPF/ArcGISRuntime.WPF.Samples/ArcGISRuntime.WPF.Samples.CSharp.csproj
@@ -56,8 +56,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.xaml.cs">
-      <DependentUpon>FeatureLayerGeoPackage.xaml</DependentUpon>
+    <Compile Include="Samples\Data\RasterLayerGeoPackage\RasterLayerGeoPackage.xaml.cs">
+      <DependentUpon>RasterLayerGeoPackage.xaml</DependentUpon>
     </Compile>
     <Compile Include="Samples\Data\AddShapefile\AddShapefile.xaml.cs">
       <DependentUpon>AddShapefile.xaml</DependentUpon>
@@ -247,7 +247,7 @@
     <Content Include="groups.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
-    <Content Include="Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.jpg">
+    <Content Include="Samples\Data\RasterLayerGeoPackage\RasterLayerGeoPackage.jpg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Samples\Data\AddShapefile\AddShapefile.jpg">
@@ -442,7 +442,7 @@
     </Content>
   </ItemGroup>
   <ItemGroup>
-    <Page Include="Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.xaml">
+    <Page Include="Samples\Data\RasterLayerGeoPackage\RasterLayerGeoPackage.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
@@ -928,7 +928,7 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
-    <Content Include="Samples\Data\FeatureLayerGeoPackage\metadata.json">
+    <Content Include="Samples\Data\RasterLayerGeoPackage\metadata.json">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
     <Content Include="Samples\Data\AddShapefile\metadata.json">
