+++ src/WPF/ArcGISRuntime.WPF.Samples/Samples/Data/RasterLayerGeoPackage/RasterLayerGeoPackage.jpg
