+++ src/WPF/ArcGISRuntime.WPF.Samples/Samples/Data/RasterLayerGeoPackage/RasterLayerGeoPackage.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl 
-    x:Class="ArcGISRuntime.WPF.Samples.FeatureLayerGeoPackage.FeatureLayerGeoPackage"
+    x:Class="ArcGISRuntime.WPF.Samples.RasterLayerGeoPackage.RasterLayerGeoPackage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013">
