+++ src/WPF/ArcGISRuntime.WPF.Samples/Samples/Data/RasterLayerGeoPackage/metadata.json
@@ -1,14 +1,14 @@
 {
-  "Name": "Feature layer (GeoPackage)",
-  "SampleName": "FeatureLayerGeoPackage",
-  "Description": "This sample demonstrates how to open a GeoPackage and show a GeoPackage feature table in a feature layer.",
+  "Name": "Raster layer (GeoPackage)",
+  "SampleName": "RasterLayerGeoPackage",
+  "Description": "This sample demonstrates how to open a GeoPackage and show a GeoPackage raster in a raster layer.",
   "Instructions": "The GeoPackage will be downloaded from an ArcGIS Online portal automatically.",
   "Type": 0,
   "RequiresOnlineConnection": true,
   "RequiresOfflineData": true,
   "DataItemIds": [ "68ec42517cdd439e81b036210483e8e7" ],
   "RequiresLocalServer": false,
-  "Image": "FeatureLayerGeoPackage.jpg",
+  "Image": "RasterLayerGeoPackage.jpg",
   "Link": "",
   "TypeLink": [ "T:Esri.ArcGISRuntime.Data.GeoPackage", "T:Esri.ArcGISRuntime.Data.GeoPackageFeatureTable", "P:Esri.ArcGISRuntime.Data.GeoPackage.GeoPackageFeatureTables", "M:Esri.ArcGISRuntime.Data.GeoPackage.OpenAsync(System.String)" ]
 }
