+++ src/iOS/Xamarin.iOS/ArcGISRuntime.Xamarin.Samples.iOS.csproj
@@ -345,7 +345,7 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </EmbeddedResource>
     <EmbeddedResource Include="Samples\Layers\TimeBasedQuery\metadata.json" />
-    <EmbeddedResource Include="Samples\Data\FeatureLayerGeoPackage\metadata.json" />
+    <EmbeddedResource Include="Samples\Data\RasterLayerGeoPackage\metadata.json" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Main.cs" />
@@ -361,7 +361,7 @@
     <Compile Include="Models\SubCategoryModel.cs" />
     <Compile Include="Samples\Data\GenerateGeodatabase\GenerateGeodatabase.cs" />
     <Compile Include="Samples\Data\FeatureLayerQuery\FeatureLayerQuery.cs" />
-    <Compile Include="Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.cs" />
+    <Compile Include="Samples\Data\RasterLayerGeoPackage\RasterLayerGeoPackage.cs" />
     <Compile Include="Samples\Data\AddShapefile\AddShapefile.cs" />
     <Compile Include="Samples\Data\ServiceFeatureTableCache\ServiceFeatureTableCache.cs" />
     <Compile Include="Samples\Data\ServiceFeatureTableManualCache\ServiceFeatureTableManualCache.cs" />
@@ -449,7 +449,7 @@
   <ItemGroup>
     <BundleResource Include="Samples\Map\ChangeBasemap\ChangeBasemap.jpg" />
     <EmbeddedResource Include="Samples\Data\AddShapefile\AddShapefile.jpg" />
-    <EmbeddedResource Include="Samples\Data\FeatureLayerGeoPackage\FeatureLayerGeoPackage.jpg" />
+    <EmbeddedResource Include="Samples\Data\RasterLayerGeoPackage\RasterLayerGeoPackage.jpg" />
     <Content Include="Samples\Data\StatisticalQuery\StatisticalQuery.jpg">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
